refactor: modularize common options for Hover Legend and Static Legend
