+++ src/visualization/components/internal/HoverLegend.tsx
@@ -1,29 +1,22 @@
 // Libraries
-import React, {ChangeEvent, CSSProperties, FC, useEffect, useState} from 'react'
+import React, {ChangeEvent, FC, useEffect, useState} from 'react'
 
 // Utils
 import {convertUserInputToNumOrNaN} from 'src/shared/utils/convertUserInput'
 
 // Components
 import {
-  AlignItems,
-  Appearance,
   ButtonShape,
   Columns,
-  ComponentColor,
-  ComponentSize,
-  FlexBox,
-  FlexDirection,
   Form,
   Grid,
-  InfluxColors,
-  InputLabel,
-  InputToggleType,
-  RangeSlider,
   SelectGroup,
-  SlideToggle,
-  Toggle,
 } from '@influxdata/clockface'
+import {
+  OrientationToggle,
+  OpacitySlider,
+  ColorizeRowsToggle,
+} from 'src/visualization/components/internal/LegendOptions'
 
 // Types
 import {VisualizationOptionProps} from 'src/visualization'
@@ -41,9 +34,6 @@ import {
 import {
   LEGEND_OPACITY_MAXIMUM,
   LEGEND_OPACITY_MINIMUM,
-  LEGEND_OPACITY_STEP,
-  LEGEND_ORIENTATION_THRESHOLD_HORIZONTAL,
-  LEGEND_ORIENTATION_THRESHOLD_VERTICAL,
   LegendDisplayStatus,
 } from 'src/visualization/constants'
 
@@ -51,7 +41,7 @@ import {
 import {event} from 'src/cloud/utils/reporting'
 
 // Styles
-import 'src/visualization/components/internal/LegendOrientation.scss'
+import 'src/visualization/components/internal/LegendOptions.scss'
 
 interface HoverLegendToggleProps {
   properties:
@@ -63,22 +53,7 @@ interface HoverLegendToggleProps {
   }
 }
 
-interface OrientationToggleProps {
-  graphType: string
-  legendOrientation: number
-  handleSetOrientation: (threshold: number) => void
-}
-
-interface OpacitySliderProps {
-  legendOpacity: number
-  handleSetOpacity: (event: ChangeEvent<HTMLInputElement>) => void
-}
-
-interface ColorizeRowsToggleProps {
-  legendColorizeRows: boolean
-  handleSetColorization: () => void
-}
-interface LegendOrientationProps extends VisualizationOptionProps {
+interface HoverLegendProps extends VisualizationOptionProps {
   properties:
     | BandViewProperties
     | XYViewProperties
@@ -90,128 +65,6 @@ interface LegendOrientationProps extends VisualizationOptionProps {
 }
 
 const eventPrefix = 'visualization.customize'
-
-const getToggleColor = (toggle: boolean): CSSProperties => {
-  if (toggle) {
-    return {color: InfluxColors.Cloud}
-  }
-  return {color: InfluxColors.Sidewalk}
-}
-
-const OrientationToggle: FC<OrientationToggleProps> = ({
-  graphType,
-  legendOrientation,
-  handleSetOrientation,
-}) => {
-  const setOrientation = (orientation: string): void => {
-    if (orientation === 'vertical') {
-      handleSetOrientation(LEGEND_ORIENTATION_THRESHOLD_VERTICAL)
-    } else {
-      handleSetOrientation(LEGEND_ORIENTATION_THRESHOLD_HORIZONTAL)
-    }
-    event(`${eventPrefix}.legend.orientation.${orientation}`, {
-      type: graphType,
-    })
-  }
-  return (
-    <FlexBox
-      direction={FlexDirection.Column}
-      margin={ComponentSize.Large}
-      alignItems={AlignItems.FlexStart}
-      className="legend-orientation-toggle"
-    >
-      <InputLabel id="legend-orientation-label">Orientation</InputLabel>
-      <Toggle
-        tabIndex={1}
-        value="horizontal"
-        name="legendOr"
-        className="legend-orientation--horizontal"
-        id="legend-orientation--horizontal"
-        checked={legendOrientation === LEGEND_ORIENTATION_THRESHOLD_HORIZONTAL}
-        onChange={setOrientation}
-        type={InputToggleType.Radio}
-        size={ComponentSize.ExtraSmall}
-        color={ComponentColor.Primary}
-        appearance={Appearance.Outline}
-      >
-        <InputLabel
-          active={legendOrientation === LEGEND_ORIENTATION_THRESHOLD_HORIZONTAL}
-          htmlFor="legend-orientation--horizontal"
-        >
-          Horizontal
-        </InputLabel>
-      </Toggle>
-      <Toggle
-        tabIndex={2}
-        value="vertical"
-        className="legend-orientation--vertical"
-        id="legend-orientation--vertical"
-        name="lengendOr"
-        checked={legendOrientation === LEGEND_ORIENTATION_THRESHOLD_VERTICAL}
-        onChange={setOrientation}
-        type={InputToggleType.Radio}
-        size={ComponentSize.ExtraSmall}
-        color={ComponentColor.Primary}
-        appearance={Appearance.Outline}
-      >
-        <InputLabel
-          active={legendOrientation === LEGEND_ORIENTATION_THRESHOLD_VERTICAL}
-          htmlFor="legend-orientation--vertical"
-        >
-          Vertical
-        </InputLabel>
-      </Toggle>
-    </FlexBox>
-  )
-}
-
-const OpacitySlider: FC<OpacitySliderProps> = ({
-  legendOpacity,
-  handleSetOpacity,
-}) => {
-  // without the toFixed(0) sometimes you
-  // can get numbers like 45.000009% which we want to avoid
-  const percentLegendOpacity = (legendOpacity * 100).toFixed(0)
-  return (
-    <Form.Element
-      className="legend-opacity-slider"
-      label={`Opacity: ${percentLegendOpacity}%`}
-    >
-      <RangeSlider
-        max={LEGEND_OPACITY_MAXIMUM}
-        min={LEGEND_OPACITY_MINIMUM}
-        step={LEGEND_OPACITY_STEP}
-        value={legendOpacity}
-        onChange={handleSetOpacity}
-        hideLabels={true}
-      />
-    </Form.Element>
-  )
-}
-
-const ColorizeRowsToggle: FC<ColorizeRowsToggleProps> = ({
-  legendColorizeRows,
-  handleSetColorization,
-}) => {
-  return (
-    <FlexBox
-      direction={FlexDirection.Row}
-      alignItems={AlignItems.Center}
-      margin={ComponentSize.Medium}
-      stretchToFitWidth={true}
-      className="legend-colorize-rows-toggle"
-    >
-      <SlideToggle
-        active={legendColorizeRows}
-        size={ComponentSize.ExtraSmall}
-        onChange={handleSetColorization}
-      />
-      <InputLabel style={getToggleColor(legendColorizeRows)}>
-        Colorize Rows
-      </InputLabel>
-    </FlexBox>
-  )
-}
 
 const HoverLegendToggle: FC<HoverLegendToggleProps> = ({
   properties,
@@ -293,10 +146,7 @@ const HoverLegendToggle: FC<HoverLegendToggleProps> = ({
   )
 }
 
-const LegendOrientation: FC<LegendOrientationProps> = ({
-  properties,
-  update,
-}) => {
+const HoverLegend: FC<HoverLegendProps> = ({properties, update}) => {
   const handleSetHoverLegendHide = (legendHide: boolean): void => {
     update({
       legendHide,
@@ -384,4 +234,4 @@ const LegendOrientation: FC<LegendOrientationProps> = ({
   )
 }
 
-export default LegendOrientation
+export default HoverLegend
