+++ src/visualization/components/internal/LegendOptions.scss
