+++ src/visualization/types/Band/options.tsx
@@ -25,7 +25,7 @@ import {
   parseYBounds,
 } from 'src/shared/utils/vis'
 import ColorSchemeDropdown from 'src/visualization/components/internal/ColorSchemeDropdown'
-import LegendOrientation from 'src/visualization/components/internal/LegendOrientation'
+import HoverLegend from 'src/visualization/components/internal/HoverLegend'
 import StaticLegend from 'src/visualization/components/internal/StaticLegend'
 import AxisTicksGenerator from 'src/visualization/components/internal/AxisTicksGenerator'
 
@@ -421,7 +421,7 @@ const BandViewOptions: FC<Props> = ({properties, results, update}) => {
           widthLG={Columns.Four}
         >
           <h5 className="view-options--header">Legend</h5>
-          <LegendOrientation
+          <HoverLegend
             properties={properties}
             results={results}
             update={update}
