+++ src/visualization/types/Histogram/options.tsx
@@ -17,7 +17,7 @@ import {
 import AutoDomainInput from 'src/shared/components/AutoDomainInput'
 import {defaultXColumn} from 'src/shared/utils/vis'
 import ColorSchemeDropdown from 'src/visualization/components/internal/ColorSchemeDropdown'
-import LegendOrientation from 'src/visualization/components/internal/LegendOrientation'
+import HoverLegend from 'src/visualization/components/internal/HoverLegend'
 import {HistogramViewProperties} from 'src/types'
 import {VisualizationOptionProps} from 'src/visualization'
 import {convertUserInputToNumOrNaN} from 'src/shared/utils/convertUserInput'
@@ -226,7 +226,7 @@ const HistogramOptions: FC<Props> = ({properties, results, update}) => {
           widthLG={Columns.Four}
         >
           <h5 className="view-options--header">Legend</h5>
-          <LegendOrientation
+          <HoverLegend
             properties={properties}
             results={results}
             update={update}
